Suppressing compilation of Xdr archive classes.
This was forced by removal of the SUN RPC library from glibc in
recent linux releases. The XdrIArchive and XdrOArchive classes
used the header <rpc/rpc.h> which used to be standard, but is no
longer available as part of the GNU standard C library, glibc.
The code for these classes remains, but they will not be compiled
by default. They should be compilable on a system with an older
glibc, or a system in which the RPC library has been explicitly
installed.